+++ src/Item/ViewItem/ViewItem.js
@@ -331,6 +331,7 @@ const ViewItem = ({
             item={item}
             tenantId={tenantTo}
             onClose={() => setIsVersionHistoryOpen(false)}
+            isInstanceShared={isInstanceShared || isInstanceShadowCopy(instance.source)}
           />
         )}
       </Paneset>